Move `fetchAppDetailsFromApiKey` inside PartnersClient
It's directly dependent on the Partners API, so it shouldn't be in the
theoretically generic services/dev/fetch.ts
