+++ packages/app/src/cli/utilities/developer-platform-client.ts
@@ -184,6 +184,20 @@ export type AssetUrlSchema = WithUserErrors<{
   assetUrl: string
 }>
 
+export enum Flag {
+  DeclarativeWebhooks,
+}
+
+const FlagMap: {[key: string]: Flag} = {
+  '5b25141b': Flag.DeclarativeWebhooks,
+}
+
+export function filterDisabledFlags(disabledFlags: string[] = []): Flag[] {
+  const defaultActiveFlags: Flag[] = [Flag.DeclarativeWebhooks]
+  const remoteDisabledFlags = disabledFlags.map((flag) => FlagMap[flag])
+  return defaultActiveFlags.filter((flag) => !remoteDisabledFlags.includes(flag))
+}
+
 export interface DeveloperPlatformClient {
   clientName: string
   webUiName: string
